SuperTrunfo nível novato em funcionamento

+++ MediaAlunos.c
@@ -5,15 +5,15 @@ e operadores de atribuição para atualizar valores. */
 # include <stdio.h>
 
 int main(){
-    int nota1, nota2, nota3;
+    float nota1, nota2, nota3;
 
     printf("Calculadora de Médias\n");
     printf("Digite a primeira nota:");
-    scanf("%d", &nota1);
+    scanf("%f", &nota1);
     printf("Digite a segunda nota:");
-    scanf("%d", &nota2);
+    scanf("%f", &nota2);
     printf("Digite a terceira nota:");
-    scanf("%d", &nota3);
+    scanf("%f", &nota3);
 
     float media = (nota1 + nota2 + nota3)/3;
